pyproject.toml
[project]
name = "py-notion-client"
version = "0.1.0"
description = "Python wrapper for Notion API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "loguru>=0.7.2",
    "pydantic-settings>=2.6.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.3",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-env>=1.1.5",
    "pytest-richer>=0.1.1",
    "ruff>=0.7.0",
    "pyright>=1.1.385",
]
package = true

[tool.pyright]
exclude = ["tests", ".venv"]


[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
asyncio_default_fixture_loop_scope = "function"
addopts = "-ra -q --rich"
testpaths = [
    "tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "ENVIRONMENT=TEST",
]

[tool.coverage.run]
omit = [
    "tests/*"
]
relative_files = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "class .*\\(.*\\bProtocol\\b.*\\):",
    "@(?:typing\\.)?overload",
    "if TYPE_CHECKING:",
    "raise NotImplementedError"
]
omit = [
    "tests/*",
]
show_missing = true
skip_covered = true
skip_empty = true
precision = 2
fail_under = 90.0
